Improve training function for neural networks
Training function now keeps a record of the training history.

Training function now periodically saves the training history, as well
as a snapshot of the model being trained.

Preprocessing of the data is now done across two separate functions: one
which is primarily performing the train-valid-test split, and one which
is local to one of the splits.