Fix bug in identifying last layer in building NN
In utils.py:get_fc_nn, the calculation of whether we were looking
at the last layer of the neural network (and hence, whether we
should use the given activation, or the final sigmoid activation)
was incorrect, and we were basically always adding the sigmoid
(as if every layer was the final layer). Since we have only built
networks with sigmoid so far, this hasn't shown up as an issue.

Either way, this is fixed now.